Update gemfile and gemspec to support solidus 4
The extension was not supported for solidus 4 with starter frontend now with this change it will
work with latest solidus version and would be conditionally using the solidus_frontend
based on the branch.

Older versions of Rails, and thus older versions of Solidus, are
incompatible with the `sqlite3` gem 1.x. This makes sure we use the
right version for each Rails version.